requirements.txt
pysimplegui
pywinctl
pillow